CanopyRadiativeTransfer

Leaf Inclination Angle

CanopyRadiativeTransfer.lidf_cdf — Function

This function returns cumulative distribution frequency. Supported methods are

lidf_cdf(lidf, θ)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/inclination_angle.jl:32.

CanopyRadiativeTransfer.lidf_cdf — Method

lidf_cdf(lidf::VerhoefLIDF{FT}, θ::FT) where {FT<:AbstractFloat}

Return the cumulative distribution frequency, given

lidf VerhoefLIDF type algorithm

θ Leaf inclination angle in [°]

CanopyRadiativeTransfer.inclination_angles! — Function

This function updates cumulative distribution frequency within the canopy. Supported methods are

inclination_angles!(can, lidf)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/inclination_angle.jl:88.

inclination_angles!(can, lidf, a, b)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/inclination_angle.jl:117.

CanopyRadiativeTransfer.inclination_angles! — Method

inclination_angles!(can::Union{BroadbandSLCanopy{FT}, HyperspectralMLCanopy{FT}}, lidf::VerhoefLIDF{FT}) where {FT<:AbstractFloat}

Update the frequency of leaf inclination angles, given

can HyperspectralMLCanopy type multiple layer canopy

lidf VerhoefLIDF type algorithm

CanopyRadiativeTransfer.inclination_angles! — Method

inclination_angles!(can::Union{BroadbandSLCanopy{FT}, HyperspectralMLCanopy{FT}}, lidf::VerhoefLIDF{FT}, a::FT, b::FT) where {FT<:AbstractFloat}

Update the frequency of leaf inclination angles, given

can BroadbandSLCanopy or HyperspectralMLCanopy type multiple layer canopy

lidf VerhoefLIDF type algorithm

a VerhoefLIDF parameter A

b VerhoefLIDF parameter B

Clumping Index

CanopyRadiativeTransfer.clumping_index! — Function

clumping_index!(can::HyperspectralMLCanopy, angles::SunSensorGeometry{FT}) where {FT<:AbstractFloat}

Update the clumping index, given

can HyperspectralMLCanopy type canopy

angles SunSensorGeometry type angles

Hyperspectral Canopy RT

CanopyRadiativeTransfer.soil_albedo! — Function

This function updates soil albedo values for shortwave PAR and NIR. The supported methods include

soil_albedo!(can, soil, albedo; clm)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/soil.jl:34.

soil_albedo!(can, soil, albedo; clm)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/soil.jl:72.

soil_albedo!(can, soil)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/soil.jl:130.

CanopyRadiativeTransfer.soil_albedo! — Method

soil_albedo!(can::HyperspectralMLCanopy{FT}, soil::Soil{FT}, albedo::BroadbandSoilAlbedo{FT}; clm::Bool = false) where {FT<:AbstractFloat}

Updates lower soil boundary reflectance, given

can HyperspectralMLCanopy type struct

soil Soil type struct

albedo BroadbandSoilAlbedo type soil albedo

CanopyRadiativeTransfer.soil_albedo! — Method

soil_albedo!(can::HyperspectralMLCanopy{FT}, soil::Soil{FT}, albedo::HyperspectralSoilAlbedo{FT}; clm::Bool = false) where {FT<:AbstractFloat}

Updates lower soil boundary reflectance, given

can HyperspectralMLCanopy type struct

soil Soil type struct

albedo HyperspectralSoilAlbedo type soil albedo

CanopyRadiativeTransfer.soil_albedo! — Method

soil_albedo!(can::HyperspectralMLCanopy{FT}, soil::Soil{FT}) where {FT<:AbstractFloat}

Updates lower soil boundary reflectance, given

can HyperspectralMLCanopy type struct

soil Soil type struct

CanopyRadiativeTransfer.extinction_coefficient — Function

This function compute the extinction coefficients for direct and diffuse radiation for given leaf inclination. Supported methods are

extinction_coefficient(sza, lia)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/coefficients.jl:32.

extinction_coefficient(lia)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/coefficients.jl:58.

extinction_coefficient(sza, vza, raa, lia)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/coefficients.jl:93.

CanopyRadiativeTransfer.extinction_coefficient — Method

extinction_coefficient(sza::FT, lia::FT) where {FT<:AbstractFloat}

Return the extinction coefficient for direct radiation, given

sza Solar zenith angle in °

lia Leaf inclination angle in °

CanopyRadiativeTransfer.extinction_coefficient — Method

extinction_coefficient(lia::FT) where {FT<:AbstractFloat}

Return the extinction coefficient for diffuse radiation (unifrom 18 average angles from 2.5° to 87.5°), given

sza Solar zenith angle in °

CanopyRadiativeTransfer.extinction_coefficient — Method

extinction_coefficient(sza::FT, vza::FT, raa::FT, lia::FT) where {FT<:AbstractFloat}

Return the extinction and scattering coefficients (extinction coefficients from solar and viewing directions, and scattering coefficients for backward and forward directions, and some sin and cos products: _Co, _Cs, _So, _Ss), given

sza Solar zenith angle in °

vza Viewing zenith angle in °

raa Relative azimuth angle in °

lia Leaf inclination angle in °

CanopyRadiativeTransfer.extinction_scattering_coefficients! — Function

This function updates the extinction (and scattering) coefficients for canopy. Supported methods are

extinction_scattering_coefficients!(can, angles)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/coefficients.jl:169.

extinction_scattering_coefficients!(can, angles)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/coefficients.jl:200.

CanopyRadiativeTransfer.extinction_scattering_coefficients! — Method

extinction_scattering_coefficients!(can::BroadbandSLCanopy{FT}, angles::SunSensorGeometry{FT}) where {FT<:AbstractFloat}

Update the extinction and scattering coefficients, given

can BroadbandSLCanopy type canopy

angles SunSensorGeometry type angles

CanopyRadiativeTransfer.extinction_scattering_coefficients! — Method

extinction_scattering_coefficients!(can::HyperspectralMLCanopy{FT}, angles::SunSensorGeometry{FT}) where {FT<:AbstractFloat}

Update the extinction and scattering coefficients, given

can HyperspectralMLCanopy type canopy

angles SunSensorGeometry type angles

CanopyRadiativeTransfer.canopy_optical_properties! — Function

This function updates canopy optical properties for canopy. The supported methods include

canopy_optical_properties!(can, angles)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/geometry.jl:38.

canopy_optical_properties!(can, albedo)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/geometry.jl:118.

canopy_optical_properties!(can, albedo)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/geometry.jl:143.

canopy_optical_properties!(can, leaves, soil)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/geometry.jl:173.

CanopyRadiativeTransfer.canopy_optical_properties! — Method

canopy_optical_properties!(can::HyperspectralMLCanopy{FT}, angles::SunSensorGeometry{FT}) where {FT<:AbstractFloat}

Updates canopy optical properties (extinction coefficients for direct and diffuse light) based on the SAIL model, given

can HyperspectralMLCanopy type struct

angles SunSensorGeometry type struct

CanopyRadiativeTransfer.canopy_optical_properties! — Method

canopy_optical_properties!(can::HyperspectralMLCanopy{FT}, leaves::Vector{Leaf{FT}}) where {FT<:AbstractFloat}

Updates canopy optical properties (scattering coefficient matrices), given

can HyperspectralMLCanopy type struct

leaves Vector of Leaf

soil Bottom soil boundary layer

CanopyRadiativeTransfer.canopy_radiation! — Function

This function updates canopy radiation profiles. The supported methods include

canopy_radiation!(can, leaf, rad, soil)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/radiation.jl:43.

canopy_radiation!(can, leaf, rad, soil)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/radiation.jl:151.

canopy_radiation!(can, albedo)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/radiation.jl:225.

canopy_radiation!(can, albedo)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/radiation.jl:257.

canopy_radiation!(can, leaves, rad, soil)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/radiation.jl:293.

canopy_radiation!(can, leaves, rad, soil)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/radiation.jl:441.

CanopyRadiativeTransfer.canopy_radiation! — Method

canopy_radiation!(can::BroadbandSLCanopy{FT}, leaf::Leaf{FT}, rad::BroadbandRadiation{FT}, soil::Soil{FT}) where {FT<:AbstractFloat}

Updates soil shortwave radiation profiles, given

can HyperspectralMLCanopy type struct

leaf Leaf type struct

rad BroadbandRadiation solar radiation

soil Soil type struct

CanopyRadiativeTransfer.canopy_radiation! — Method

canopy_radiation!(can::BroadbandSLCanopy{FT}, leaf::Leaf{FT}, rad::FT, soil::Soil{FT}) where {FT<:AbstractFloat}

Updates soil longwave radiation profiles, given

can HyperspectralMLCanopy type struct

leaf Leaf type struct

rad Incoming longwave radiation

soil Soil type struct

CanopyRadiativeTransfer.canopy_radiation! — Method

canopy_radiation!(can::HyperspectralMLCanopy{FT}, albedo::BroadbandSoilAlbedo{FT}) where {FT<:AbstractFloat}

Updates soil shortwave radiation profiles, given

can HyperspectralMLCanopy type struct

albedo BroadbandSoilAlbedo type soil albedo

CanopyRadiativeTransfer.canopy_radiation! — Method

canopy_radiation!(can::HyperspectralMLCanopy{FT}, albedo::HyperspectralSoilAlbedo{FT}) where {FT<:AbstractFloat}

Updates soil shortwave radiation profiles, given

can HyperspectralMLCanopy type struct

albedo HyperspectralSoilAlbedo type soil albedo

CanopyRadiativeTransfer.canopy_radiation! — Method

canopy_radiation!(can::HyperspectralMLCanopy{FT}, leaves::Vector{Leaf{FT}}, rad::HyperspectralRadiation{FT}, soil::Soil{FT}) where {FT<:AbstractFloat}

Updates canopy radiation profiles for shortwave radiation, given

can HyperspectralMLCanopy type struct

leaves Vector of Leaf

rad Incoming solar radiation

soil Bottom soil boundary layer

CanopyRadiativeTransfer.canopy_radiation! — Method

canopy_radiation!(can::HyperspectralMLCanopy{FT}, leaves::Vector{Leaf{FT}}, rad::FT, soil::Soil{FT}) where {FT<:AbstractFloat}

Updates canopy radiation profiles for longwave radiation, given

can HyperspectralMLCanopy type struct

leaves Vector of Leaf

rad Incoming longwave radiation

soil Bottom soil boundary layer

CanopyRadiativeTransfer.canopy_fluorescence! — Function

This function updates canopy fluorescence profiles. The supported methods include

canopy_fluorescence!(can, leaves; ϕ_photon)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/fluorescence.jl:38.

CanopyRadiativeTransfer.canopy_fluorescence! — Method

canopy_fluorescence!(can::HyperspectralMLCanopy{FT}; ϕ_photon::Bool = true) where {FT<:AbstractFloat}

Updates canopy radiation profiles for shortwave radiation, given

can HyperspectralMLCanopy type struct

leaves Vector of Leaf

ϕ_photon If true (default), convert photon to photon when computing SIF; otherwise, convert energy to energy

Remote Sensing Applications

CanopyRadiativeTransfer.read_spectrum — Function

This function interpolate the spectrum to give values at the target wavelength bin(s). The supported methods include

read_spectrum(x, y, target)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/remote_sensing.jl:33.

read_spectrum(x, y, x₁, x₂; steps)

defined at /home/runner/work/CanopyRadiativeTransfer.jl/CanopyRadiativeTransfer.jl/src/remote_sensing.jl:68.

CanopyRadiativeTransfer.read_spectrum — Method

read_spectrum(x::Vector{FT}, y::Vector{FT}, target::FT) where {FT<:AbstractFloat}

Return the spectrum value at target wavelength bin, given

x X-axis of the spectrum

y Y-axis of the spectrum

target Target x value

CanopyRadiativeTransfer.read_spectrum — Method

read_spectrum(x::Vector{FT}, y::Vector{FT}, x₁::FT, x₂::FT; steps::Int = 2) where {FT<:AbstractFloat}

Return the spectrum value at target wavelength bin, given

x X-axis of the spectrum

y Y-axis of the spectrum

x₁ Lower x boundary

x₂ Upper x boundary

steps The incremental Δx is (x₂ - x₁) / steps

CanopyRadiativeTransfer.MODIS_EVI — Function

MODIS_EVI(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return EVI for MODIS setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.MODIS_EVI2 — Function

MODIS_EVI2(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return EVI2 for MODIS setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.MODIS_LSWI — Function

MODIS_LSWI(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return LSWI for MODIS setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.MODIS_NDVI — Function

MODIS_NDVI(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return NDVI for MODIS setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.MODIS_NIRv — Function

MODIS_NIRv(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return NIRv for MODIS setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.OCO2_SIF759 — Function

OCO2_SIF759(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return SIF @ 759 nm for OCO2 setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.OCO2_SIF770 — Function

OCO2_SIF770(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return SIF @ 770 nm for OCO2 setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.OCO3_SIF759 — Function

OCO3_SIF759(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return SIF @ 759 nm for OCO3 setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.OCO3_SIF770 — Function

OCO3_SIF770(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return SIF @ 770 nm for OCO3 setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.TROPOMI_SIF683 — Function

TROPOMI_SIF683(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return SIF @ 682.5 nm for TROPOMI setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.TROPOMI_SIF740 — Function

TROPOMI_SIF740(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return SIF @ 740 nm for TROPOMI setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.TROPOMI_SIF747 — Function

TROPOMI_SIF747(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return SIF @ 746.5 nm for TROPOMI setup, given

can HyperspectralMLCanopy type canopy

CanopyRadiativeTransfer.TROPOMI_SIF751 — Function

TROPOMI_SIF751(can::HyperspectralMLCanopy{FT}) where {FT<:AbstractFloat}

Return SIF @ 750.5 nm for TROPOMI setup, given

can HyperspectralMLCanopy type canopy